+++ src/test/java/com/cschar/pmode3/GSONJUnit4Test.java
@@ -1,0 +1,20 @@
+package com.cschar.pmode3;
+
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+//Dummy test for pipeline
+public class GSONJUnit4Test {
+    @Test
+    public void testSerializeColor() {
+        ZStateBasicParticle z = new ZStateBasicParticle();
+        ZStateBasicParticleConverter c = new ZStateBasicParticleConverter();
+
+        ZStateBasicParticle zDeserialized = c.fromString(c.toString(z));
+
+        assertEquals(z.basicColor.getRGB(), zDeserialized.basicColor.getRGB());
+    }
+
+}